Move map panel components to a different folder
