+++ app/src/components/MapPanels/MiniGlobe.jsx
