+++ app/src/components/MapPanels/leftControlPanel/components/Loader.jsx
