+++ app/src/components/MapPanels/leftControlPanel/components/ZoomControls.jsx
