+++ app/src/components/MapPanels/leftControlPanel/containers/Loader.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import Loader from 'leftControlPanel/components/Loader';
+import Loader from 'components/MapPanels/leftControlPanel/components/Loader';
 
 const mapStateToProps = state => ({
   visible: state.map.loading || (state.vesselInfo.infoPanelStatus && state.vesselInfo.infoPanelStatus.isLoading)
