+++ app/src/components/MapPanels/rightControlPanel/containers/SubMenu.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
-import SubMenu from 'rightControlPanel/components/SubMenu';
-import { setSubmenu } from 'rightControlPanel/rightControlPanelActions';
+import SubMenu from 'components/MapPanels/rightControlPanel/components/SubMenu';
+import { setSubmenu } from 'components/MapPanels/rightControlPanel/rightControlPanelActions';
 
 
 const mapDispatchToProps = dispatch => ({
