+++ app/src/components/MapPanels/rightControlPanel/rightControlPanelActions.js
